improvements

+++ matrix_clock/drivers/sercom/twi.h
@@ -30,10 +30,10 @@ void twi_off(void);
 void twi_on(void);
 
 void twi_stop_transmission(TWI_t* const TWI);
-uint8_t twi_start_transmission(TWI_t* const TWI, const uint8_t slave_address, const uint8_t timeout_ms);
-uint8_t twi_send_byte(TWI_t* const TWI, const uint8_t timeout_ms, const uint8_t data);
-uint8_t twi_receive_byte(TWI_t* const TWI, const uint8_t timeout_ms, uint8_t* const data, const uint8_t end_of_data);
-uint8_t twi_read_packet(TWI_t* const TWI, const uint8_t slave_address, const uint8_t timeout_ms, const uint8_t reg, uint8_t* data, uint8_t length);
-uint8_t twi_write_packet(TWI_t* const TWI, const uint8_t slave_address, const uint8_t timeout_ms, const uint8_t reg, const uint8_t* data, uint8_t length);
+uint8_t twi_start_transmission(TWI_t* const TWI, const uint8_t slave_address, const uint16_t timeout_ms);
+uint8_t twi_send_byte(TWI_t* const TWI, const uint16_t timeout_ms, const uint8_t data);
+uint8_t twi_receive_byte(TWI_t* const TWI, const uint16_t timeout_ms, uint8_t* const data, const uint8_t end_of_data);
+uint8_t twi_read_packet(TWI_t* const TWI, const uint8_t slave_address, const uint16_t timeout_ms, const uint8_t reg, uint8_t* data, uint8_t length);
+uint8_t twi_write_packet(TWI_t* const TWI, const uint8_t slave_address, const uint16_t timeout_ms, const uint8_t reg, const uint8_t* data, uint8_t length);
 
 #endif /* TWI_H_ */
